service/pyproject.toml
[tool.poetry]
name = "properties-prices-register"
version = "0.1.0"
description = ""
authors = ["Lucio Tudisco <lucio.tudisco@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.0.1"
django-celery-results = "^2.5.1"
python-dotenv = "^1.0.0"
psycopg2-binary = "^2.9.9"
requests = "^2.31.0"
types-requests = "^2.31.0.20231231"
djangorestframework = "^3.14.0"
markdown = "^3.5.1"
django-filter = "^23.5"
algoliasearch = "^3.0.0"
boto3 = "^1.34.12"
sentry-sdk = {extras = ["flask"], version = "1.39.1"}
django-health-check = "^3.17.0"
gunicorn = "^21.2.0"
celery = "^5.3.6"
eventlet = "^0.34.2"
drf-aggregation = "^1.1.2"

[tool.poetry.group.dev.dependencies]
pre-commit = "^3.6.0"
localstack = "^3.0.2"
awscli = "^1.32.14"
awscli-local = "^0.22.0"
pytest = "^7.4.4"
pytest-cov = "^4.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
